+++ packages/BiostructureViewer/src/tests/pdbqt-tests.ts
@@ -2,20 +2,19 @@ import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
 import * as ui from 'datagrok-api/ui';
 
-import wu from 'wu';
 import * as ngl from 'NGL';
 
 import {category, expect/*, expect*/, expectObject, test} from '@datagrok-libraries/utils/src/test';
 import {Molecule3DUnitsHandler} from '@datagrok-libraries/bio/src/molecule-3d';
-
-import {Pdbqt} from '../utils/pdbqt/pdbqt-parser';
 import {errInfo} from '@datagrok-libraries/bio/src/utils/err-info';
-import {importPdbqt} from '../package';
-import {IPdbAtomBase, IPdbqtAtomBase} from '../utils/pdbqt/types';
-import {AtomBase, AtomCoordsBase, LineBase} from '../utils/pdbqt/types-base';
-import {PdbqtAtomCoords} from '../utils/pdbqt/types-pdbqt';
-import {PdbAtomCoords} from '../utils/pdbqt/types-pdb';
+import {IPdbAtomBase, IPdbqtAtomBase} from '@datagrok-libraries/bio/src/pdb/format/types';
+import {AtomBase, AtomCoordsBase, LineBase} from '@datagrok-libraries/bio/src/pdb/format/types-base';
+import {PdbAtomCoords} from '@datagrok-libraries/bio/src/pdb/format/types-pdb';
+
 import {IMPORT} from '../consts-import';
+import {importPdbqtUI} from '../utils/import-pdbqt';
+import {PdbqtAtomCoords} from '../utils/types-pdbqt';
+import {Pdbqt} from '../utils/pdbqt-parser';
 
 import {_package} from '../package-test';
 
@@ -173,7 +172,7 @@ async function _testPdbqtParseAutodockGpu(): Promise<void> {
 async function _testPdbqtImportModels(): Promise<void> {
   try {
     const cnt: string = await _package.files.readAsText('docking/ligand_out.pdbqt');
-    const df: DG.DataFrame = (await importPdbqt(cnt, true))[0];
+    const df: DG.DataFrame = (await importPdbqtUI(cnt, true))[0];
 
     const molCol = df.getCol(IMPORT.pdb.molColName);
     const uh = IMPORT.pdb.unitsHandlerClass.getOrCreate(molCol);
@@ -187,7 +186,7 @@ async function _testPdbqtImportModels(): Promise<void> {
 
 async function _testPdbqtImportTargetOnly(): Promise<void> {
   const cnt: string = await _package.files.readAsText('docking/3SWZ.pdbqt');
-  const dfList = await importPdbqt(cnt, true);
+  const dfList = await importPdbqtUI(cnt, true);
   expect(dfList.length, 0);
 
   const view = grok.shell.v;
@@ -218,4 +217,3 @@ async function _testNglPdbqtParserLigands(): Promise<void> {
   expect(pose0Val.atomCount, 30);
   expect(pose0Val.bondCount, 34);
 }
-